Fix bugs, improve code quality, and add CMake build system
- Fix localplayer: use dynamic index instead of hardcoded 1, prevent stale pointers
- Fix static init order: defer interface/netvar init to DllMain after modules load
- Fix entity null checks: remove UB this==nullptr patterns
- Fix entity loops: start from index 1 (skip world entity)
- Fix mouse tracer condition: && to || for AlwaysOn logic
- Remove unreachable edge-bug detection in auto-strafer
- Slim hook.h includes, move them to implementation files
- Replace norm macro with truncf_to_int inline function
- Fix PlaySound macro conflict: rename to playSoundFile
- Fix compiler warnings (srand cast, tellg type, printf format)
- Improve AnimationLib: smooth switch transitions, O(1) easing lookup
- Add CMake build system (replaces .vcxproj)
- Update loader DLL path for CMake output
- Extract SDK headers (EngineClient, ClientClasses, PlayerInfo, etc.)

+++ CSGO_v22/CSGO_v2/CSGO_v2/entity.h
@@ -2,7 +2,7 @@
 #include "netvars.h"
 #include <iostream>
 #include <vector>
-#include "math.h"
+#include "GameMath.h"
 
 #define STR_MERGE_IMPL(a, b) a##b
 #define STR_MERGE(a, b) STR_MERGE_IMPL(a,b)
@@ -12,7 +12,7 @@
 #define NETVAR_DECL(name, type, offset) \
     type name() noexcept \
 {\
-    return *(type*)(this + offset); \
+    return *(type*)((uintptr_t)this + offset); \
 }
 
 // bounding box
@@ -34,7 +34,7 @@ struct BBox {
     math::Vector brt; // Back Right Top
 
     // Dimensions
-    int w, h;
+    float w, h;
 
     bool isValid = false;
 };
